requirements-dev.txt
Flask>=0.7
Flask-SQLAlchemy>=0.15
peewee
wtf-peewee
mongoengine<0.11.0
pymongo==2.8
flask-mongoengine==0.8.2
pillow==2.9.0
Babel<=1.3
flask-babelex
shapely==1.5.9
geoalchemy2
psycopg2
nose
sphinx
sphinx-intl
coveralls
pylint
python-nmap
